+++ ca-gateway/scep-client/src/test/java/org/xipki/scep/client/test/CaTest.java
@@ -14,11 +14,6 @@ import org.xipki.scep.transaction.CaCapability;
 public class CaTest extends AbstractCaTest {
 
   @Override
-  protected boolean isWithRa() {
-    return true;
-  }
-
-  @Override
   protected CaCapability[] getExcludedCaCaps() {
     return null;
   }